refactors to provide for additional texts through separate modules - See issue #10 for discussion
