+++ scriptures/texts/deuterocanon.py
@@ -1,0 +1,26 @@
+from __future__ import unicode_literals
+from .base import Text
+
+class Deuterocanon(Text):
+    """
+    Deuterocanonical Books
+    """
+    books = [
+        ('Tobit', 'Tob', 'Tob(?:it)?', [22, 14, 17, 21, 22, 17, 18, 21, 6, 12, 19, 22, 18, 15,]),
+
+        # TODO:
+        # Judith
+        # Additions to Esther
+        # Wisdom of Solomon
+        # Ecclesiasticus (aka Sirach)
+        # Baruch
+        # Letter of Jeremiah
+        # Prayer of Azariah
+        # Susanna
+        # Bel and the Dragon
+        # Prayer of Manasseh
+        # 1 Maccabees
+        # 2 Maccabees
+
+    ]
+
